Cambio de int a long en Alumno

+++ src/main/java/cua/integra/cajasuamc/entity/Alumno.java
@@ -17,10 +17,10 @@ import java.util.List;
 public class Alumno {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
 
     @Column(nullable = false, unique = true)
-    private int matricula;
+    private long matricula;
 
     @Column(nullable = false)
     private String nombre;
